pyproject.toml
[tool.poetry]
name = "user-management-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.96.0"
passlib = "^1.7.4"
pydantic = "^1.10.9"
SQLAlchemy = "^2.0.15"
python-multipart = "^0.0.6"
uvicorn = "^0.22.0"
psycopg2-binary = "^2.9.6"
black = "^23.3.0"
isort = "^5.12.0"
bcrypt = "^4.0.1"
async-asgi-testclient = "^1.4.11"
pytest = "^7.3.2"
pytest-asyncio = "^0.21.0"
docker = "^6.1.3"
pdbpp = "^0.10.3"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
